Tests/LibWeb/Text/expected/Worker/SharedWorker-close-one-owner.txt

worker2: connect 2
worker2: echo 2: after-gc
